Realice las correcciones y agregue los 2 querys

+++ src/main/java/com/upc/campusflow/Repository/RecursoRepository.java
@@ -1,5 +1,5 @@
 package com.upc.campusflow.Repository;
 import com.upc.campusflow.Model.Recurso;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface RecursoRepository extends JpaRepository<Recurso, Integer> {
+public interface RecursoRepository extends JpaRepository<Recurso, Long> {
 }
